:recycle: Introduce abstract prediction module
